move applyMiddleware -> applyBatchMiddleware for batchedNetworkInterface
