+++ src/transport/networkInterface.ts
@@ -18,23 +18,6 @@ import {
   isBatchAfterwares,
 } from './afterware';
 
-import {
-  BatchRequestAndOptions,
-  BatchResponseAndOptions,
-} from './batchedNetworkInterface';
-
-function isBatchRequestOptions(
-  requestAndOptions: BatchRequestAndOptions | RequestAndOptions,
-): requestAndOptions is BatchRequestAndOptions {
-  return (<BatchRequestAndOptions>requestAndOptions).requests && !(<RequestAndOptions>requestAndOptions).request;
-}
-
-function isBatchResponseOptions(
-  responseAndOptions: BatchResponseAndOptions | ResponseAndOptions,
-): responseAndOptions is BatchResponseAndOptions {
-  return (<BatchResponseAndOptions>responseAndOptions).responses && !(<ResponseAndOptions>responseAndOptions).response;
-}
-
 /**
  * This is an interface that describes an GraphQL document to be sent
  * to the server.
@@ -109,8 +92,6 @@ export function printRequest(request: Request): PrintedRequest {
   };
 }
 
-// TODO: refactor
-// add the batching to this.
 export class HTTPFetchNetworkInterface implements NetworkInterface {
   public _uri: string;
   public _opts: RequestInit;
@@ -136,100 +117,50 @@ export class HTTPFetchNetworkInterface implements NetworkInterface {
     this._batchAfterwares = [];
   }
 
-  public applyMiddlewares(requestAndOptions: BatchRequestAndOptions): Promise<BatchRequestAndOptions>
-  public applyMiddlewares(requestAndOptions: RequestAndOptions): Promise<RequestAndOptions>
-  public applyMiddlewares(
-    requestAndOptions: RequestAndOptions | BatchRequestAndOptions,
-  ): Promise<RequestAndOptions | BatchRequestAndOptions> {
+  public applyMiddlewares(requestAndOptions: RequestAndOptions): Promise<RequestAndOptions> {
     return new Promise((resolve, reject) => {
-      if (isBatchRequestOptions(requestAndOptions)) {
-        const { requests, options } = requestAndOptions;
-        const queue = (funcs: BatchMiddlewareInterface[], scope: any) => {
-          const next = () => {
-            if (funcs.length > 0) {
-              const f = funcs.shift();
-              if (f) {
-                f.applyBatchMiddleware.apply(scope, [{ requests, options }, next]);
-              }
-            } else {
-              resolve({
-                requests,
-                options,
-              });
+      const { request, options } = requestAndOptions;
+      const queue = (funcs: MiddlewareInterface[], scope: any) => {
+        const next = () => {
+          if (funcs.length > 0) {
+            const f = funcs.shift();
+            if (f) {
+              f.applyMiddleware.apply(scope, [{ request, options }, next]);
             }
-          };
-          next();
+          } else {
+            resolve({
+              request,
+              options,
+            });
+          }
         };
+        next();
+      };
 
-        queue([...this._batchMiddlewares], this);
-      } else {
-        const { request, options } = requestAndOptions;
-        const queue = (funcs: MiddlewareInterface[], scope: any) => {
-          const next = () => {
-            if (funcs.length > 0) {
-              const f = funcs.shift();
-              if (f) {
-                f.applyMiddleware.apply(scope, [{ request, options }, next]);
-              }
-            } else {
-              resolve({
-                request,
-                options,
-              });
-            }
-          };
-          next();
-        };
-
-        queue([...this._middlewares], this);
-      }
+      queue([...this._middlewares], this);
     });
   }
 
-  public applyAfterwares(responseAndOptions: ResponseAndOptions): Promise<ResponseAndOptions>
-  public applyAfterwares(responseAndOptions: BatchResponseAndOptions): Promise<BatchResponseAndOptions>
-  public applyAfterwares(
-    responseAndOptions: ResponseAndOptions | BatchResponseAndOptions,
-  ): Promise<ResponseAndOptions | BatchResponseAndOptions> {
+  public applyAfterwares({response, options}: ResponseAndOptions): Promise<ResponseAndOptions> {
     return new Promise((resolve, reject) => {
       // Declare responseObject so that afterware can mutate it.
-      if (isBatchResponseOptions(responseAndOptions)) {
-        const responseObject = responseAndOptions;
-        const queue = (funcs: BatchAfterwareInterface[], scope: any) => {
-          const next = () => {
-            if (funcs.length > 0) {
-              const f = funcs.shift();
-              if (f) {
-                f.applyBatchAfterware.apply(scope, [responseObject, next]);
-              }
-            } else {
-              resolve(responseObject);
+      const responseObject = {response, options};
+      const queue = (funcs: AfterwareInterface[], scope: any) => {
+        const next = () => {
+          if (funcs.length > 0) {
+            const f = funcs.shift();
+            if (f) {
+              f.applyAfterware.apply(scope, [responseObject, next]);
             }
-          };
-          next();
+          } else {
+            resolve(responseObject);
+          }
         };
+        next();
+      };
 
-        // iterate through afterwares using next callback
-        queue([...this._batchAfterwares], this);
-      } else {
-        const responseObject = responseAndOptions;
-        const queue = (funcs: AfterwareInterface[], scope: any) => {
-          const next = () => {
-            if (funcs.length > 0) {
-              const f = funcs.shift();
-              if (f) {
-                f.applyAfterware.apply(scope, [responseObject, next]);
-              }
-            } else {
-              resolve(responseObject);
-            }
-          };
-          next();
-        };
-
-        // iterate through afterwares using next callback
-        queue([...this._afterwares], this);
-      }
+      // iterate through afterwares using next callback
+      queue([...this._afterwares], this);
     });
   }
 
@@ -300,7 +231,7 @@ export class HTTPFetchNetworkInterface implements NetworkInterface {
         if (typeof middleware.applyBatchMiddleware === 'function') {
           this._batchMiddlewares.push(middleware);
         } else {
-          throw new Error('Middleware must implement the applyMiddleware function');
+          throw new Error('Middleware must implement the applyBatchMiddleware function');
         }
       });
     }
@@ -323,7 +254,7 @@ export class HTTPFetchNetworkInterface implements NetworkInterface {
         if (typeof afterware.applyBatchAfterware === 'function') {
           this._batchAfterwares.push(afterware);
         } else {
-          throw new Error('Afterware must implement the applyAfterware function');
+          throw new Error('Afterware must implement the applyBatchAfterware function');
         }
       });
     }
